functional
+ issue found, at any exception doesn't make a rollback, test it!

+++ application/apis/models/borrow_model.py
@@ -1,6 +1,4 @@
-from tokenize import Double
-
-from sqlalchemy import Column, Integer, Boolean, DateTime
+from sqlalchemy import Column, Integer, Boolean, DateTime, Float
 from sqlalchemy.ext.declarative import declarative_base
 
 Base = declarative_base()
@@ -10,9 +8,9 @@ class Borrows(Base):
     __tablename__ = "borrows"
 
     idborrow = Column(Integer, primary_key=True, index=True)
-    borrowamount = Column(Double)
-    percentagetax = Column(Double)
-    totalpayment = Column(Double)
+    borrowamount = Column(Float)
+    percentagetax = Column(Float)
+    totalpayment = Column(Float)
     datetaken = Column(DateTime)
     duedate = Column(DateTime)
     idmethod = Column(Integer)
