+++ application/apis/models/issued_bill_model.py
@@ -1,6 +1,4 @@
-from tokenize import Double
-
-from sqlalchemy import Column, Integer, String, Boolean, DateTime
+from sqlalchemy import Column, Integer, DateTime, Float
 from sqlalchemy.ext.declarative import declarative_base
 
 Base = declarative_base()
@@ -11,7 +9,7 @@ class IssuedBill(Base):
 
     idbill = Column(Integer, primary_key=True, index=True)
     duedate = Column(DateTime)
-    amounttopay = Column(Double)
+    amounttopay = Column(Float)
     idplan = Column(Integer)
 
     def __str__(self):
